Fix: Add is_admin field to UserResponse API model

+++ backend/alembic/versions/2026_01_27_add_birthday_wall_archive_fields.py
@@ -21,9 +21,10 @@ def upgrade():
     op.add_column('birthday_walls', sa.Column('birthday_year', sa.Integer(), nullable=True))
     
     # Set birthday_year for existing walls based on opens_at date
+    # opens_at is 24h before birthday, so birthday is opens_at + 1 day
     op.execute("""
         UPDATE birthday_walls 
-        SET birthday_year = EXTRACT(YEAR FROM opens_at + INTERVAL '1 day')
+        SET birthday_year = EXTRACT(YEAR FROM opens_at + INTERVAL '1 day')::INTEGER
         WHERE birthday_year IS NULL
     """)
